+++ src/color-highlight.js
@@ -8,7 +8,7 @@ import { findScssVars } from './strategies/scss-vars';
 import { findLessVars } from './strategies/less-vars';
 import { findStylVars } from './strategies/styl-vars';
 import { findCssVars } from './strategies/css-vars';
-import { findFn } from './strategies/functions';
+import { findColorFunctionsInText } from './strategies/functions';
 import { findRgbNoFn } from './strategies/rgbWithoutFunction';
 import { findHexARGB, findHexRGBA } from './strategies/hex';
 import { findHwb } from './strategies/hwb';
@@ -31,7 +31,7 @@ export class DocumentHighlight {
     this.disposed = false;
 
     this.document = document;
-    this.strategies = [findFn, findHwb];
+    this.strategies = [findColorFunctionsInText, findHwb];
 
     if (viewConfig.useARGB == true) {
       this.strategies.push(findHexARGB);