+++ src/strategies/scss-vars.js
@@ -1,6 +1,6 @@
 import { findHexRGBA } from './hex';
 import { findWords } from './words';
-import { findFn, sortStringsDesc } from './functions';
+import { findColorFunctionsInText, sortStringsInDescendingOrder } from './functions';
 import { findHwb } from './hwb';
 import { parseImports } from '../lib/sass-importer';
 
@@ -36,7 +36,7 @@ export async function findScssVars(text, importerOptions) {
     const values = await Promise.race([
       findHexRGBA(value),
       findWords(value),
-      findFn(value),
+      findColorFunctionsInText(value),
       findHwb(value)
     ]);
 
@@ -52,7 +52,7 @@ export async function findScssVars(text, importerOptions) {
     return [];
   }
 
-  varNames = sortStringsDesc(varNames);
+  varNames = sortStringsInDescendingOrder(varNames);
 
   const varNamesRegex = new RegExp(`\\$(${varNames.join('|')})(?!-|\\s*:)`, 'g');